Undrop admin db

+++ core/services/mongodb-remote/src/test/java/com/mongodb/stitch/core/services/mongodb/remote/sync/internal/SyncUnitTestHarness.kt
@@ -712,7 +712,9 @@ class SyncUnitTestHarness : Closeable {
 
     override fun close() {
         latestCtx?.localClient?.listDatabaseNames()?.forEach {
-            latestCtx?.localClient?.getDatabase(it)?.drop()
+            if (it != "admin") {
+                latestCtx?.localClient?.getDatabase(it)?.drop()
+            }
         }
         latestCtx?.dataSynchronizer?.close()
     }
